[Refactor] #163 - 작품 검색 Data 구현을 NovelData에서 SearchData로 이동
