+++ Projects/Data/SearchData/Sources/DTO/Query/DetailSearchQuery.swift
@@ -1,6 +1,6 @@
 //
 //  DetailSearchQuery.swift
-//  NovelData
+//  SearchData
 //
 //  Created by Seoyeon Choi on 3/27/26.
 //  Copyright © 2026 kr.websoso.app. All rights reserved.
