+++ Projects/Data/SearchData/Sources/DTO/Response/SearchNovelsResponse.swift
@@ -1,6 +1,6 @@
 //
 //  SearchNovelsResponse.swift
-//  NovelData
+//  SearchData
 //
 //  Created by Seoyeon Choi on 3/27/26.
 //  Copyright © 2026 kr.websoso.app. All rights reserved.
