+++ Projects/Data/SearchData/Sources/Endpoint/SearchEndpoint.swift
@@ -15,7 +15,9 @@ enum SearchEndpoint: Endpoint {
     case deleteRecentSearchWord(id: Int)
     case deleteAllRecentSearchWords
     case getAutoCompletionWords(SearchAutoCompletionQuery)
-    
+    case getNormalSearchResult(NormalSearchQuery)
+    case getDetailSearchResult(DetailSearchQuery)
+
     var baseURL: URL {
         URL(string: NetworkingConfig.baseURL) ?? URL(string: "")!
     }
@@ -28,12 +30,16 @@ enum SearchEndpoint: Endpoint {
             return "/novels/recent-searches/\(id)"
         case .getAutoCompletionWords:
             return "/novels/autocomplete"
+        case .getNormalSearchResult:
+            return "/novels"
+        case .getDetailSearchResult:
+            return "/novels/filtered"
         }
     }
 
     var method: HTTPMethod {
         switch self {
-        case .getRecentSearchWords, .getAutoCompletionWords:
+        case .getRecentSearchWords, .getAutoCompletionWords, .getNormalSearchResult, .getDetailSearchResult:
             return .get
         case .deleteRecentSearchWord, .deleteAllRecentSearchWords:
             return .delete
@@ -43,6 +49,10 @@ enum SearchEndpoint: Endpoint {
     var query: QueryParameters {
         switch self {
         case .getAutoCompletionWords(let query):
+            return .convertible(query)
+        case .getNormalSearchResult(let query):
+            return .convertible(query)
+        case .getDetailSearchResult(let query):
             return .convertible(query)
         case .getRecentSearchWords, .deleteRecentSearchWord, .deleteAllRecentSearchWords:
             return .none
@@ -55,10 +65,15 @@ enum SearchEndpoint: Endpoint {
 
     var authorization: AuthorizationPolicy {
         switch self {
-        case .getRecentSearchWords, .deleteRecentSearchWord, .deleteAllRecentSearchWords:
+        case .getRecentSearchWords,
+                .deleteRecentSearchWord,
+                .deleteAllRecentSearchWords,
+                .getDetailSearchResult:
             return .requireToken
         case .getAutoCompletionWords:
             return .usesTokenIfAvailable
+        case .getNormalSearchResult:
+            return .withoutToken
         }
     }
 }
