+++ Projects/Data/SearchData/Tests/SearchDataTests.swift
@@ -1,0 +1,6 @@
+//
+//  SearchDataTests.swift
+//  SearchDataTests
+//
+
+import Testing
